added downgrade check for tables if they exist before dropping them t…

+++ migrations/versions/198fe7ae3fc9_create_customer_table.py
@@ -7,6 +7,7 @@ Create Date: 2020-11-16 13:06:58.250956
 """
 from alembic import op
 import sqlalchemy as sa
+from sqlalchemy.engine.reflection import Inspector
 
 
 # revision identifiers, used by Alembic.
@@ -35,4 +36,8 @@ def upgrade():
 
 
 def downgrade():
-    op.drop_table('customers')
+    conn = op.get_bind()
+    inspector = Inspector.from_engine(conn)
+    tables = inspector.get_table_names()
+    if 'customers' in tables:
+        op.drop_table('customers')
